Move distributions into random_dist.h/cpp
