+++ CMakeLists.txt
@@ -303,6 +303,7 @@ list(APPEND libopenmc_SOURCES
   src/plot.cpp
   src/position.cpp
   src/progress_bar.cpp
+  src/random_dist.cpp
   src/random_lcg.cpp
   src/reaction.cpp
   src/reaction_product.cpp
